Use community version of Tensorflow that works with M1
The TensorFlow binary downloaded from a normal TensorFlow 2.3.1 pip install (from requirements)
was crashing when we used the linux/x86_64 emulated arch with M1 macs (which is needed
because TensorFlow does not yet have an arm-supported version).

To solve this, we are using a community wheel of Tensorflow 2.3.1 compiled as we need it.

More on this here: https://github.com/tensorflow/tensorflow/issues/52845

Paired with Ahmed!

CHECK-2147