CV2-2805 (#284)
* CV2-2805 Fix support for minimum elastic search score

This commit fixes a bug where the `min_es_score` parameter was silently being discarded.
It adds a test for the expected behavior and fixes the bug.
It also removes redundant code. We were removing vectors in both restrict_results and in strip_vectors. I removed the code from the former.