Changing the value of `body['models']` to contain only successfully retrieved models' vectors. So that the item gets saved in ElasticSearch successfully if OpenAI API fails
If OpenAI API fails and the openai- model vector is null, we want it to fail successfully. In this case, we want it ES to continue saving the item but removing the OpenAI vector. This fix commit is to change the value of `body['models']` to remove `openai-*` in case it failed. Otherwise, the item doesn't get saved.