upgrade redis to 6.2 and specify travis distro
We had to ugrade Redis in Pender due to the Sidekiq upgrade, and Redis
is on 6.2 on our qa and live environment. So we are taking this opportunity
to update in all repos.

Regarding the distro, we had issues running the tests on Travis,
in Pender changing the disto fixed it. So I'm trying that here as well.