Trying to fix build error on check repo (#505)
I noticed that the `check` build was installing a different version of `tokenizer`, despite the pinned version in `requirements.txt` (maybe it was coming from another dependency). That different version was triggering an error on `check`'s CI. This PR ensures that the version we need is installed.

I also took the opportunity to fix some failing tests.

Reference: CV2-6381.